+++ govtool/backend/src/VVA/Survey.hs
@@ -2,10 +2,10 @@
 {-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE TemplateHaskell   #-}
 
-module VVA.Account where
+module VVA.Survey where
 
 import           Control.Monad.Except        (MonadError, throwError)
-import           Control.Monad.Reader        (MonadIO, MonadReader, liftIO)
+import           Control.Monad.Reader
 import           Control.Monad.Trans.Control (MonadBaseControl)
 
 import           Data.ByteString             (ByteString)
@@ -14,26 +14,25 @@ import           Data.Has                    (Has)
 import           Data.String                 (fromString)
 import           Data.Text                   (Text, unpack)
 import qualified Data.Text.Encoding          as Text
-import qualified Data.Text.IO                as Text
 
 import qualified Database.PostgreSQL.Simple  as SQL
 
 import           VVA.Pool                    (ConnectionPool, withPool)
-import           VVA.Types                   (AccountInfo (..), AppError (..))
+import           VVA.Types                   (AppError (..))
 
 sqlFrom :: ByteString -> SQL.Query
-sqlFrom = fromString . unpack . Text.decodeUtf8
+sqlFrom bs = fromString $ unpack $ Text.decodeUtf8 bs
 
-accountInfoSql :: SQL.Query
-accountInfoSql = sqlFrom $(embedFile "sql/get-account-info.sql")
+getSurveyDefinitionSql :: SQL.Query
+getSurveyDefinitionSql = sqlFrom $(embedFile "sql/get-survey-definition.sql")
 
-accountInfo ::
+getSurveyDefinition ::
   (Has ConnectionPool r, MonadReader r m, MonadIO m, MonadBaseControl IO m, MonadError AppError m) =>
   Text ->
-    m AccountInfo
-accountInfo stakeKey = withPool $ \conn -> do
-  result <- liftIO $ SQL.query conn accountInfoSql (SQL.Only stakeKey)
+  m Text
+getSurveyDefinition txId = withPool $ \conn -> do
+  result <- liftIO $ SQL.query conn getSurveyDefinitionSql (SQL.Only txId)
   case result of
-    [(id, view, is_registered, is_script_based)] ->
-      return $ AccountInfo id view is_registered is_script_based
-    _ -> throwError $ CriticalError "Could not query the account info."
+    [SQL.Only payload] -> pure payload
+    _ -> throwError $ NotFoundError $
+      "No metadata label 17 found for transaction " <> txId